// src/app/modules/broadcomproxy_module/pages/wizards/broadcom-proxy/broadcom-proxy.component.ts
import { ChangeDetectionStrategy, Component, inject, ViewChild } from '@angular/core';
import { WizardsAbstractComponent } from '../../../../../pages/wizards/wizards-abstract/wizards-abstract.component';
import { SelectKeyValueString } from '../../../../../layouts/primeng/select.interface';
import {
    WizardsDynamicfieldsComponent
} from '../../../../../components/wizards/wizards-dynamicfields/wizards-dynamicfields.component';
import { RouterLink } from '@angular/router';
import { FaIconComponent } from '@fortawesome/angular-fontawesome';
import {
    CardBodyComponent,
    CardComponent,
    CardHeaderComponent,
    CardTitleDirective,
    FormControlDirective,
    FormLabelDirective
} from '@coreui/angular';
import { TranslocoDirective, TranslocoPipe } from '@jsverse/transloco';
import { RequiredIconComponent } from '../../../../../components/required-icon/required-icon.component';
import { BackButtonDirective } from '../../../../../directives/back-button.directive';
import { BroadcomProxyWizardPost } from './broadcom-proxy-wizard.interface';
import { BroadcomProxyWizardService } from './broadcom-proxy-wizard.service';
import { SelectComponent } from '../../../../../layouts/primeng/select/select/select.component';
import { FormErrorDirective } from '../../../../../layouts/coreui/form-error.directive';
import { FormFeedbackComponent } from '../../../../../layouts/coreui/form-feedback/form-feedback.component';
import { FormsModule } from '@angular/forms';


@Component({
    selector: 'oitc-broadcom-proxy',
    imports: [
    RouterLink,
    FaIconComponent,
    CardComponent,
    CardHeaderComponent,
    CardBodyComponent,
    BackButtonDirective,
    TranslocoPipe,
    RequiredIconComponent,
    FormLabelDirective,
    SelectComponent,
    FormErrorDirective,
    FormFeedbackComponent,
    FormsModule,
    FormControlDirective,
    WizardsDynamicfieldsComponent,
    TranslocoDirective,
    CardTitleDirective
],
    templateUrl: './broadcom-proxy.component.html',
    styleUrl: './broadcom-proxy.component.css',
    changeDetection: ChangeDetectionStrategy.OnPush
})
export class BroadcomProxyComponent extends WizardsAbstractComponent {
    @ViewChild(WizardsDynamicfieldsComponent) childComponentLocal!: WizardsDynamicfieldsComponent;
    protected override WizardService: BroadcomProxyWizardService = inject(BroadcomProxyWizardService);
    public checked: boolean = false;

    protected override post: BroadcomProxyWizardPost = {
// Default fields from the base wizard
        host_id: 0,
        services: [],
// Fields for the wizard
        authPassword: '',
        authProtocol: 'md5',
        privacyPassword: '',
        privacyProtocol: 'des',
        securityLevel: '1',
        securityName: '',
        snmpCommunity: '',
        snmpVersion: '2'
    } as BroadcomProxyWizardPost;
    protected snmpVersions: SelectKeyValueString[] = [
        {value: '1', key: 'SNMP V 1'},
        {value: '2', key: 'SNMP V 2c'},
        {value: '3', key: 'SNMP V 3'},
    ]


    protected securityLevels: SelectKeyValueString[] = [
        {key: 'authPriv', value: '1'},
        {key: 'authNoPriv', value: '2'},
        {key: 'noAuthNoPriv', value: '3'},
    ];
    protected authProtocols: SelectKeyValueString[] = [
        {key: 'MD5', value: 'md5'},
        {key: 'SHA', value: 'sha'},
    ];
    protected privacyProtocols: SelectKeyValueString[] = [
        {key: 'DES', value: 'des'},
        {key: 'AES', value: 'aes'},
        {key: 'AES128', value: 'aes128'},
        {key: '3DES', value: '3des'},
        {key: '3DESDE', value: '3desde'},
    ];

}

